requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0